chore(http): split plain_client.rs out of ext-http lib.rs + move new FFI decls to stdlib_ffi_part2 (2000-line lint gate)
