+++ crates/perry-codegen/src/runtime_decls/stdlib_ffi_part2.rs
@@ -12,4 +12,13 @@ pub(crate) fn declare_stdlib_ffi_part2(module: &mut LlModule) {
     module.declare_function("js_node_http_server_unref", I64, &[I64]);
     module.declare_function("js_node_https_server_ref", I64, &[I64]);
     module.declare_function("js_node_https_server_unref", I64, &[I64]);
+    // Streaming-bodies PR — client flushHeaders early dispatch + numeric
+    // httpVersion halves.
+    module.declare_function(
+        "js_http_client_request_flush_headers",
+        DOUBLE,
+        &[I64, DOUBLE, DOUBLE],
+    );
+    module.declare_function("js_node_http_im_http_version_major", DOUBLE, &[I64]);
+    module.declare_function("js_node_http_im_http_version_minor", DOUBLE, &[I64]);
 }
